[Refactor] #7 과도한 추상화 방지하기 위해 Buttons 파일 제거, DialogButton만 별도로 타입 관리
